Rename directories in plural
