+++ src/components/builder/grimoires/PanelGrimoire.tsx
